* [core]: add CIE*xy diagram from EDID values and JSON details
The device infos are shown delayed to allow noticing of changes.

The CIE*xy diagram layout is choosen very carefully to make sense
in many sizes and window ratios. The canvas2d text is scaled to
match the QML font sizes.
The CIE*xy is grayed out while showing the JSON keys for better
readability.

The button for Show more details is placed on the device page, as
it makes only sense in this context. Details are reset each
program start to have one consistent state.